Contribution How To Guide.md
Contributing to Dual Sensor Sensation(Group 24): 

 

Designers for the HAT: 

 

Tlokomelo Nlosi:NKSREG002@myuct.ac.za 

Kabelo Manake:MNKKAB001@myuct.ac.za 

Shane Napoleon:NPLSHA003@myuct.ac.za 

 

Contributions to this project are welcome, however, the guide below must be read and adhered to 

 

 

1)How To Contribute 

 

First, email one of the designers (Tlokomelo Nkosi, Kabelo Manake or Shane Napoleon), detailing what is wished to change/contribute. Once approval has been received from one of the designers, submit a pull request. A branch will be created in GitLab where contributions and changes can be worked on, once the changes/contributions are complete, submit a merge request and the contributions will be evaluated and either used or discarded. 

 

2) Finding Issues (Repository, Documentation, PCB) 

 

If issues are found in the GitLab repository, be it to do with the PCB, Documentation or even just the way the repository is structured, please raise an issue on GitHub or email one of the designers so that they can take a look at the problem. 

 

3) Extending Repository, Contributing to the PCB/Code 

 

When contributing to the repository, please ensure that all contributions are in well-written English, that is in a well set out document format. 

When contributing to the PCB, only KiCAD software may be used, alongside KiCADs standard library. If there is a footprint or anything that is missing from the KiCAD library, please seek approval by emailing one of the designers, detailing what you would like to use and from which library you will use it from. 

When contributing to the code please ensure that it is well commented and follows the same format that the original code was written in. The code compiler/editer that must be used is STM32CubeIDE. 

 

4) Lastly, ensure that all changes are well documented and easy to understand.